+++ simulator/src/main.rs
@@ -10,8 +10,14 @@ use rdkafka::{
     util::Timeout,
 };
 use runs::{
-    create_alarm_command, create_run_start_command, create_run_stop_command, create_runlog_command,
-    create_sample_environment_command, AlarmData, RunLogData, SampleEnvData, Start, Stop,
+    create_messages::{
+        create_alarm_command,
+        create_run_start_command,
+        create_run_stop_command,
+        create_runlog_command,
+        create_sample_environment_command
+    },
+    AlarmData, RunLogData, SampleEnvData, Start, Stop,
 };
 use std::{
     path::PathBuf,
